Implement Repo.store()
Use ETS (or DETS) to store invoices.

Repo.start_link() takes a factory function for creating the table
rather than a table itself so that the Repo process owns the table.
Otherwise, the calling process would have to create the table with
`:public` access, which may not be desirable.

Also implement Repo.find() to aid in testing.