⬆️ Upgrade to tzdata 1.0.2
Up until now, I've been holding off on this upgrade because newer versions were trying to download an updated timezone database, which didn't work in an escript (see https://github.com/lau/tzdata/issues/24).

But now, it is possible to turn off auto-updates and configure a different data directory, which solves the issue.